requirements.txt
django
djangorestframework
django-environ
psycopg2
djangorestframework-simplejwt